move workload_fio_storageutilization into fiojob
